+++ src/main/java/org/osiam/resources/scim/Email.java
@@ -63,6 +63,7 @@ public final class Email extends MultiValuedAttribute implements Serializable {
         this.type = builder.type;
     }
 
+    @Deprecated
     @Override
     public String getOperation() {
         return super.getOperation();
@@ -158,6 +159,7 @@ public final class Email extends MultiValuedAttribute implements Serializable {
             type = email.type;
         }
 
+        @Deprecated
         @Override
         public Builder setOperation(String operation) {
             super.setOperation(operation);